Refactor the QC process
