+++ PopPUNK/__main__.py
@@ -34,8 +34,12 @@ def get_options():
             help='Create pairwise distances database between reference sequences',
             default=False,
             action='store_true')
+    mode.add_argument('--qc-db',
+            help='Run quality control on a reference database',
+            default=False,
+            action='store_true')
     mode.add_argument('--fit-model',
-            help='Fit a mixture model to a reference database',
+            help='Fit a model to a (QCed) reference database',
             choices=['bgmm', 'dbscan', 'refine', 'lineage', 'threshold'],
             default = False)
     mode.add_argument('--use-model',
@@ -77,15 +81,18 @@ def get_options():
 
     # qc options
     qcGroup = parser.add_argument_group('Quality control options')
-    qcGroup.add_argument('--qc-filter', help='Behaviour following sequence QC step: "stop" [default], "prune"'
-                                                ' (analyse data passing QC), or "continue" (analyse all data)',
-                                                default='stop', type = str, choices=['stop', 'prune', 'continue'])
+    qcGroup.add_argument('--qc-keep', help="Only write failing sequences to a file, don't remove them from the database file",
+                                      default=False, action='store_true')
+    qcGroup.add_argument('--remove-samples', help='A list of names to remove from the database (regardless of any other QC)',
+                                     default=None, type=str)
     qcGroup.add_argument('--retain-failures', help='Retain sketches of genomes that do not pass QC filters in '
-                                                'separate database [default = False]', default=False, action='store_true')
+                                                   'separate database [default = False]', default=False, action='store_true')
     qcGroup.add_argument('--max-a-dist', help='Maximum accessory distance to permit [default = 0.5]',
                                                 default = 0.5, type = float)
     qcGroup.add_argument('--max-pi-dist', help='Maximum core distance to permit [default = 0.5]',
                                                 default = 0.5, type = float)
+    qcGroup.add_argument('--max-zero-dist', help='Maximum proportion of zero distances to permit [default = 0.05]',
+                                                default = 0.05, type = float)
     qcGroup.add_argument('--type-isolate', help='Isolate from which distances will be calculated for pruning [default = None]',
                                                 default = None, type = str)
     qcGroup.add_argument('--length-sigma', help='Number of standard deviations of length distribution beyond '
@@ -198,8 +205,7 @@ def main():
 
     # Imports are here because graph tool is very slow to load
     from .models import loadClusterFit, BGMMFit, DBSCANFit, RefineFit, LineageFit
-    from .sketchlib import checkSketchlibLibrary
-    from .sketchlib import removeFromDB
+    from .sketchlib import checkSketchlibLibrary, removeFromDB
 
     from .network import construct_network_from_edge_list
     from .network import construct_network_from_assignments
@@ -211,12 +217,11 @@ def main():
     from .plot import writeClusterCsv
     from .plot import plot_scatter
 
-    from .prune_db import prune_distance_matrix
+    from .qc import prune_distance_matrix, qcDistMat, sketchlibAssemblyQC, remove_qc_fail
 
     from .utils import setGtThreads
     from .utils import setupDBFuncs
     from .utils import readPickle
-    from .utils import qcDistMat
     from .utils import createOverallLineage
 
     # check kmer properties
@@ -231,22 +236,10 @@ def main():
         sys.exit(1)
     kmers = np.arange(args.min_k, args.max_k + 1, args.k_step)
 
-    # Dict of QC options for passing to database construction and querying functions
-    qc_dict = {
-        'run_qc': args.create_db,
-        'qc_filter': args.qc_filter,
-        'retain_failures': args.retain_failures,
-        'length_sigma': args.length_sigma,
-        'length_range': args.length_range,
-        'prop_n': args.prop_n,
-        'upper_n': args.upper_n,
-        'max_pi_dist': args.max_pi_dist,
-        'max_a_dist': args.max_a_dist,
-        'type_isolate': args.type_isolate
-    }
+
 
     # Dict of DB access functions
-    dbFuncs = setupDBFuncs(args, qc_dict)
+    dbFuncs = setupDBFuncs(args)
     createDatabaseDir = dbFuncs['createDatabaseDir']
     constructDatabase = dbFuncs['constructDatabase']
     queryDatabase = dbFuncs['queryDatabase']
@@ -262,13 +255,6 @@ def main():
         sys.exit(1)
     # for sketchlib, only supporting a single sketch size
     sketch_sizes = int(round(max(sketch_sizes.values())/64))
-
-    # if a length range is specified, check it makes sense
-    if args.length_range[0] is not None:
-        if args.length_range[0] >= args.length_range[1]:
-            sys.stderr.write('Ensure the specified length range is space-separated argument of'
-                             ' length 2, with the lower value first\n')
-            sys.exit(1)
 
     # check if working with lineages
     if args.fit_model == 'lineage':
@@ -299,20 +285,19 @@ def main():
 
         # generate sketches and QC sequences to identify sequences not matching specified criteria
         createDatabaseDir(args.output, kmers)
-        seq_names_passing = \
-            constructDatabase(
-                args.r_files,
-                kmers,
-                sketch_sizes,
-                args.output,
-                args.threads,
-                args.overwrite,
-                codon_phased = args.codon_phased,
-                calc_random = True)
+        seq_names = constructDatabase(
+            args.r_files,
+            kmers,
+            sketch_sizes,
+            args.output,
+            args.threads,
+            args.overwrite,
+            codon_phased = args.codon_phased,
+            calc_random = True)
 
         # calculate distances between sequences
-        distMat = queryDatabase(rNames = seq_names_passing,
-                                qNames = seq_names_passing,
+        distMat = queryDatabase(rNames = seq_names,
+                                qNames = seq_names,
                                 dbPrefix = args.output,
                                 queryPrefix = args.output,
                                 klist = kmers,
@@ -320,19 +305,82 @@ def main():
                                 number_plot_fits = args.plot_fit,
                                 threads = args.threads)
 
-        # QC pairwise distances to identify long distances indicative of anomalous sequences in the collection
-        seq_names_passing, distMat = qcDistMat(distMat,
-                                                seq_names_passing,
-                                                seq_names_passing,
-                                                args.output,
-                                                args.output,
-                                                qc_dict)
-
         # Plot results
         if not args.no_plot:
             plot_scatter(distMat,
                          args.output + "/" + os.path.basename(args.output) + "_distanceDistribution",
                          args.output + " distances")
+
+    #******************************#
+    #*                            *#
+    #* DB QC                      *#
+    #*                            *#
+    #******************************#
+    if args.qc_db:
+        # if a length range is specified, check it makes sense
+        if args.length_range[0] is not None:
+            if args.length_range[0] >= args.length_range[1]:
+                sys.stderr.write('Ensure the specified length range is space-separated argument of'
+                                ' length 2, with the lower value first\n')
+                sys.exit(1)
+
+        # Dict of QC options for passing to database construction and querying functions
+        qc_dict = {
+            'no_remove': args.qc_keep,
+            'retain_failures': args.retain_failures,
+            'length_sigma': args.length_sigma,
+            'length_range': args.length_range,
+            'prop_n': args.prop_n,
+            'upper_n': args.upper_n,
+            'max_pi_dist': args.max_pi_dist,
+            'max_a_dist': args.max_a_dist,
+            'prop_zero': args.max_zero_dist,
+            'type_isolate': args.type_isolate
+        }
+
+        refList, queryList, self, distMat = readPickle(distances, enforce_self=True)
+
+        fail_unconditionally = {}
+        # Unconditional removal
+        if args.remove_samples:
+            with open(args.remove_samples, 'r') as f:
+                for line in f:
+                    fail_unconditionally[line.rstrip] = ["removed"]
+
+        # assembly qc
+        sys.stderr.write("Running sequence QC\n")
+        pass_assembly_qc, fail_assembly_qc = \
+            sketchlibAssemblyQC(args.output,
+                                refList,
+                                qc_dict)
+        sys.stderr.write(f"{len(fail_assembly_qc)} samples failed\n")
+
+        # QC pairwise distances to identify long distances indicative of anomalous sequences in the collection
+        sys.stderr.write("Running distance QC\n")
+        pass_dist_qc, fail_dist_qc = \
+            qcDistMat(distMat,
+                      refList,
+                      queryList,
+                      args.ref_db,
+                      output,
+                      qc_dict)
+        sys.stderr.write(f"{len(fail_dist_qc)} samples failed\n")
+
+        # Get list of passing samples
+        pass_list = set(refList) - fail_unconditionally.keys() - fail_assembly_qc.keys() - fail_dist_qc.keys()
+        assert(pass_list == set(refList).intersection(set(pass_assembly_qc)).intersection(set(pass_dist_qc)))
+        passed = [x for x in refList if x not in pass_list]
+        if qc_dict['type_isolate'] is not None and qc_dict['type_isolate'] not in pass_list:
+            raise RuntimeError('Type isolate ' + qc_dict['type_isolate'] + \
+                               ' not found in isolates after QC; check '
+                               'name of type isolate and QC options\n')
+
+        if len(passed) < len(refList):
+            remove_qc_fail(qc_dict, refList, passed,
+                           [fail_unconditionally, fail_assembly_qc, fail_dist_qc],
+                           args.ref_db, distMat, args.output,
+                           args.strand_preserved, args.threads)
+
 
     #******************************#
     #*                            *#
@@ -379,11 +427,6 @@ def main():
 
         # Load the distances
         refList, queryList, self, distMat = readPickle(distances, enforce_self=True)
-        seq_names = set(set(refList) | set(queryList))
-        seq_names_passing, distMat = qcDistMat(distMat, refList, queryList, args.ref_db, output, qc_dict)
-        if len(set(seq_names_passing).difference(seq_names)) > 0 and args.qc_filter == "stop":
-            sys.stderr.write("Distances failed quality control (change QC options to run anyway)\n")
-            sys.exit(1)
 
         #******************************#
         #*                            *#
