--- poppunk_prune-runner.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-# vim: set fileencoding=<utf-8> :
-# Copyright 2018-2020 John Lees and Nick Croucher
-
-"""Convenience wrapper for running prune_poppunk directly from source tree."""
-
-from PopPUNK.prune_db import main
-
-if __name__ == '__main__':
-    main()
